% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_factor.R
\name{rc_factor}
\alias{rc_factor}
\title{Recodes a character vector into a factor}
\usage{
rc_factor(x, from, to, level_idx = NULL)
}
\arguments{
\item{x}{character vector that will be recoded and turned into a factor}

\item{from}{character vector containing each unique value in \code{x}}

\item{to}{character vector containing new values that will be used
as a substitute for values in \code{from}}

\item{level_idx}{The order in which \code{to} will be incorporated into the
new factor}
}
\value{
A factor
}
\description{
Recodes values of \code{x} based on pairs of values in \code{from} and
\code{to}, outputting a factor. Optionally allows the ordering of factor
levels based on \code{level_idx}.
}
\examples{
x <- as.character(1:10)
from <- as.character(1:10)
to <- letters[1:10]
level_idx <- rev(1:10)
recode(x, from, to, level_idx)
}
