% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recodes exact matches in a character string}
\usage{
recode(x, from, to, warn = TRUE, default_NA = FALSE)
}
\arguments{
\item{x}{A vector containing values that will be recoded based on
\code{from} and \code{to}}

\item{from}{A vector containing the values within \code{x} that will
be recoded}

\item{to}{A vector containing the corresponding value (or values) that
\code{from} will be converted to}

\item{warn}{If 'TRUE', a warning will be generated when values in 'x'
are not contained in 'from}

\item{default_NA}{If 'TRUE', \code{recode} will convert values from \code{x}
not in \code{from} to \code{NA}}
}
\value{
A character vector, derived from \code{x}, with 0 or more values
substituted for new ones
}
\description{
Finds values in \code{x} that are equivalent to values in
\code{from} and substitutes them with corresponding values in \code{to}.
}
\details{
If only a single value is included in \code{to}, then all values in
\code{x} matching
values in \code{from} will be changed to the single value in \code{to}.
Otherwise, the lengths of \code{from} and \code{to} must be equal, and there
can be no duplicated values in \code{from}.

Note that when anticipating output of \code{recode}, this function
automatically converts all input arguments to character vectors in the
first step.
}
\examples{
x <- as.character(1:10)
from <- as.character(1:10)
to <- letters[1:10]
recode(x, from, to)
}
