% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_factor.R
\name{rc_factor}
\alias{rc_factor}
\title{Recodes a character vector into a factor}
\usage{
rc_factor(x, from, to, level_idx = NULL)
}
\arguments{
\item{x}{vector that will be recoded and turned into a factor}

\item{from}{vector containing all unique values in \code{x}}

\item{to}{vector containing values that will replace those in \code{from} -
should be ordered in a manner so that each value in \code{to} occurs in this
vector as the corresponding value in \code{from}}

\item{level_idx}{vector specifying the sequence of levels in
\code{rc_factor}'s output - position of each value should be indexed to
corresponding value in \code{to} and \code{from}.}
}
\value{
A vector or class \code{factor}
}
\description{
Recodes values of \code{x} based on pairs of values in \code{from} and
\code{to}, outputting a factor. Optionally allows the ordering of factor
levels based on \code{level_idx}.
}
\examples{
x <- as.character(1:10)
from <- as.character(1:10)
to <- letters[1:10]
level_idx <- rev(1:10)
rc_char(x, from, to, level_idx)
}
