% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recodes exact matches in a character string}
\usage{
recode(x, from, to)
}
\arguments{
\item{x}{A character vector containing values that will be recoded based on
\code{from} and \code{to}}

\item{from}{A character vector containing the values within \code{x} that will
be recoded}

\item{to}{The corresponding value (or values) that \code{from} will be
converted to}
}
\value{
A character vector, derived from \code{x}, with 0 or more values
substituted for new ones
}
\description{
Finds values in \code{x} that are equivalent to values in
\code{from} and substitutes them with corresponding values in \code{to}. If only
a single value is included in \code{to}, then all values in \code{x} matching
values in \code{from} will be changed to the single value in \code{to}.
Otherwise, the lengths of \code{from} and \code{to} must be equal, and there
can be no duplicated values in \code{from}.
}
\examples{
x <- as.character(1:10)
from <- as.character(1:10)
to <- letters[1:10]
recode(x, from, to)
}
